string repeat

+++ stars/src/lib.rs
@@ -1,3 +1,3 @@
 pub fn stars(n: u32) -> String {
-"*".to_string().repeat(n)
+"*".to_string().repeat(n as usize)
 }
